+++ modules/Scheduling/includes/Courses.fnc.php
@@ -711,13 +711,18 @@ function CoursePeriodUpdateTeacher( $cp_id, $old_teacher_id, $new_teacher_id )
 
 	// Update gradebook_assignment_types.
 	// Fix SQL update gradebook_assignment_types only for old teacher's assignments
+	// Or if old teacher has no more course periods in this course
 	DBQuery( "UPDATE gradebook_assignment_types
 		SET STAFF_ID='" . (int) $new_teacher_id . "'
 		WHERE STAFF_ID='" . (int) $old_teacher_id . "'
-		AND ASSIGNMENT_TYPE_ID IN(SELECT ASSIGNMENT_TYPE_ID
-			FROM gradebook_assignments
-			WHERE STAFF_ID='" . (int) $old_teacher_id . "'
-			AND COURSE_PERIOD_ID='" . (int) $cp_id . "')" );
+		AND (ASSIGNMENT_TYPE_ID IN(SELECT ASSIGNMENT_TYPE_ID
+				FROM gradebook_assignments
+				WHERE STAFF_ID='" . (int) $old_teacher_id . "'
+				AND COURSE_PERIOD_ID='" . (int) $cp_id . "')
+			OR NOT EXISTS(SELECT 1 FROM course_periods
+				WHERE COURSE_ID=(SELECT COURSE_ID
+					WHERE COURSE_PERIOD_ID='" . (int) $cp_id . "')
+				AND TEACHER_ID='" . (int) $old_teacher_id . "'))" );
 
 	// Update gradebook_assignments.
 	DBUpdate(