+++ CHANGES.md
@@ -16,6 +16,7 @@ Changes in 11.5
 Changes in 11.4.3
 -----------------
 - Fix SQL update gradebook_assignment_types only for old teacher's assignments in Courses.fnc.php
+- Fix SQL Get all assignment types having assignments for this course period in GradebookBreakdown.php, Assignments.php, Grades.php & Courses.fnc.php
 
 Changes in 11.4.2
 -----------------